Tpetra: refactor MatrixMatrix
