+++ packages/tpetra/core/ext/TpetraExt_MatrixMatrix_decl.hpp
@@ -464,27 +464,27 @@ template <class Scalar,
           class Node,
           class LocalOrdinalViewType>
 struct KernelWrappers {
-  static inline void mult_A_B_newmatrix_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
-                                                       CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
-                                                       const LocalOrdinalViewType& Acol2Brow,
-                                                       const LocalOrdinalViewType& Acol2Irow,
-                                                       const LocalOrdinalViewType& Bcol2Ccol,
-                                                       const LocalOrdinalViewType& Icol2Ccol,
-                                                       CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
-                                                       Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Cimport,
-                                                       const std::string& label                           = std::string(),
-                                                       const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
+  static void mult_A_B_newmatrix_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
+                                                CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
+                                                const LocalOrdinalViewType& Acol2Brow,
+                                                const LocalOrdinalViewType& Acol2Irow,
+                                                const LocalOrdinalViewType& Bcol2Ccol,
+                                                const LocalOrdinalViewType& Icol2Ccol,
+                                                CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
+                                                Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Cimport,
+                                                const std::string& label                           = std::string(),
+                                                const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
 
-  static inline void mult_A_B_reuse_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
-                                                   CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
-                                                   const LocalOrdinalViewType& Acol2Brow,
-                                                   const LocalOrdinalViewType& Acol2Irow,
-                                                   const LocalOrdinalViewType& Bcol2Ccol,
-                                                   const LocalOrdinalViewType& Icol2Ccol,
-                                                   CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
-                                                   Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Cimport,
-                                                   const std::string& label                           = std::string(),
-                                                   const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
+  static void mult_A_B_reuse_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
+                                            CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
+                                            const LocalOrdinalViewType& Acol2Brow,
+                                            const LocalOrdinalViewType& Acol2Irow,
+                                            const LocalOrdinalViewType& Bcol2Ccol,
+                                            const LocalOrdinalViewType& Icol2Ccol,
+                                            CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
+                                            Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Cimport,
+                                            const std::string& label                           = std::string(),
+                                            const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
 };
 
 // Jacobi Kernel wrappers struct
@@ -496,31 +496,44 @@ template <class Scalar,
           class Node,
           class LocalOrdinalViewType>
 struct KernelWrappers2 {
-  static inline void jacobi_A_B_newmatrix_kernel_wrapper(Scalar omega,
-                                                         const Vector<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Dinv,
-                                                         CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
-                                                         CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
-                                                         const LocalOrdinalViewType& Acol2Brow,
-                                                         const LocalOrdinalViewType& Acol2Irow,
-                                                         const LocalOrdinalViewType& Bcol2Ccol,
-                                                         const LocalOrdinalViewType& Icol2Ccol,
-                                                         CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
-                                                         Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Cimport,
-                                                         const std::string& label                           = std::string(),
-                                                         const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
+  static void jacobi_A_B_newmatrix_kernel_wrapper(Scalar omega,
+                                                  const Vector<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Dinv,
+                                                  CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
+                                                  CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
+                                                  const LocalOrdinalViewType& Acol2Brow,
+                                                  const LocalOrdinalViewType& Acol2Irow,
+                                                  const LocalOrdinalViewType& Bcol2Ccol,
+                                                  const LocalOrdinalViewType& Icol2Ccol,
+                                                  CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
+                                                  Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Cimport,
+                                                  const std::string& label                           = std::string(),
+                                                  const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
 
-  static inline void jacobi_A_B_reuse_kernel_wrapper(Scalar omega,
-                                                     const Vector<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Dinv,
-                                                     CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
-                                                     CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
-                                                     const LocalOrdinalViewType& Acol2Brow,
-                                                     const LocalOrdinalViewType& Acol2Irow,
-                                                     const LocalOrdinalViewType& Bcol2Ccol,
-                                                     const LocalOrdinalViewType& Icol2Ccol,
-                                                     CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
-                                                     Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Cimport,
-                                                     const std::string& label                           = std::string(),
-                                                     const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
+  static void jacobi_A_B_reuse_kernel_wrapper(Scalar omega,
+                                              const Vector<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Dinv,
+                                              CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
+                                              CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
+                                              const LocalOrdinalViewType& Acol2Brow,
+                                              const LocalOrdinalViewType& Acol2Irow,
+                                              const LocalOrdinalViewType& Bcol2Ccol,
+                                              const LocalOrdinalViewType& Icol2Ccol,
+                                              CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
+                                              Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Cimport,
+                                              const std::string& label                           = std::string(),
+                                              const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
+
+  static void jacobi_A_B_newmatrix_KokkosKernels(Scalar omega,
+                                                 const Vector<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Dinv,
+                                                 CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
+                                                 CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
+                                                 const LocalOrdinalViewType& Acol2Brow,
+                                                 const LocalOrdinalViewType& Acol2Irow,
+                                                 const LocalOrdinalViewType& Bcol2Ccol,
+                                                 const LocalOrdinalViewType& Icol2Ccol,
+                                                 CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
+                                                 Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Cimport,
+                                                 const std::string& label                           = std::string(),
+                                                 const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
 };
 
 // Triple-Product Kernel wrappers struct
@@ -531,51 +544,51 @@ template <class Scalar,
           class Node,
           class LocalOrdinalViewType>
 struct KernelWrappers3 {
-  static inline void mult_R_A_P_newmatrix_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Rview,
-                                                         CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
-                                                         CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Pview,
-                                                         const LocalOrdinalViewType& Acol2Prow,
-                                                         const LocalOrdinalViewType& Acol2PIrow,
-                                                         const LocalOrdinalViewType& Pcol2Ccol,
-                                                         const LocalOrdinalViewType& PIcol2Ccol,
-                                                         CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Ac,
-                                                         Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Acimport,
-                                                         const std::string& label                           = std::string(),
-                                                         const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
+  static void mult_R_A_P_newmatrix_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Rview,
+                                                  CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
+                                                  CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Pview,
+                                                  const LocalOrdinalViewType& Acol2Prow,
+                                                  const LocalOrdinalViewType& Acol2PIrow,
+                                                  const LocalOrdinalViewType& Pcol2Ccol,
+                                                  const LocalOrdinalViewType& PIcol2Ccol,
+                                                  CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Ac,
+                                                  Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Acimport,
+                                                  const std::string& label                           = std::string(),
+                                                  const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
 
-  static inline void mult_R_A_P_reuse_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Rview,
-                                                     CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
-                                                     CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Pview,
-                                                     const LocalOrdinalViewType& Acol2Prow,
-                                                     const LocalOrdinalViewType& Acol2PIrow,
-                                                     const LocalOrdinalViewType& Pcol2Ccol,
-                                                     const LocalOrdinalViewType& PIcol2Ccol,
-                                                     CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Ac,
-                                                     Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Acimport,
-                                                     const std::string& label                           = std::string(),
-                                                     const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
+  static void mult_R_A_P_reuse_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Rview,
+                                              CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
+                                              CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Pview,
+                                              const LocalOrdinalViewType& Acol2Prow,
+                                              const LocalOrdinalViewType& Acol2PIrow,
+                                              const LocalOrdinalViewType& Pcol2Ccol,
+                                              const LocalOrdinalViewType& PIcol2Ccol,
+                                              CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Ac,
+                                              Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Acimport,
+                                              const std::string& label                           = std::string(),
+                                              const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
 
-  static inline void mult_PT_A_P_newmatrix_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
-                                                          CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Pview,
-                                                          const LocalOrdinalViewType& Acol2Prow,
-                                                          const LocalOrdinalViewType& Acol2PIrow,
-                                                          const LocalOrdinalViewType& Pcol2Ccol,
-                                                          const LocalOrdinalViewType& PIcol2Ccol,
-                                                          CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Ac,
-                                                          Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Acimport,
-                                                          const std::string& label                           = std::string(),
-                                                          const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
+  static void mult_PT_A_P_newmatrix_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
+                                                   CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Pview,
+                                                   const LocalOrdinalViewType& Acol2Prow,
+                                                   const LocalOrdinalViewType& Acol2PIrow,
+                                                   const LocalOrdinalViewType& Pcol2Ccol,
+                                                   const LocalOrdinalViewType& PIcol2Ccol,
+                                                   CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Ac,
+                                                   Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Acimport,
+                                                   const std::string& label                           = std::string(),
+                                                   const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
 
-  static inline void mult_PT_A_P_reuse_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
-                                                      CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Pview,
-                                                      const LocalOrdinalViewType& Acol2Prow,
-                                                      const LocalOrdinalViewType& Acol2PIrow,
-                                                      const LocalOrdinalViewType& Pcol2Ccol,
-                                                      const LocalOrdinalViewType& PIcol2Ccol,
-                                                      CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Ac,
-                                                      Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Acimport,
-                                                      const std::string& label                           = std::string(),
-                                                      const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
+  static void mult_PT_A_P_reuse_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
+                                               CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Pview,
+                                               const LocalOrdinalViewType& Acol2Prow,
+                                               const LocalOrdinalViewType& Acol2PIrow,
+                                               const LocalOrdinalViewType& Pcol2Ccol,
+                                               const LocalOrdinalViewType& PIcol2Ccol,
+                                               CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Ac,
+                                               Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node> > Acimport,
+                                               const std::string& label                           = std::string(),
+                                               const Teuchos::RCP<Teuchos::ParameterList>& params = Teuchos::null);
 };
 
 // This only merges matrices that look like B & Bimport, aka, they have no overlapping rows
