+++ packages/tpetra/core/ext/TpetraExt_MatrixMatrix_def.hpp
@@ -38,6 +38,8 @@
 #include "TpetraExt_MatrixMatrix_ExtraKernels_def.hpp"
 #include "Tpetra_Details_getEntryOnHost.hpp"
 
+#include "TpetraExt_MatrixMatrix_KernelWrappers.hpp"
+
 #include "KokkosSparse_spgemm.hpp"
 #include "KokkosSparse_spadd.hpp"
 #include "Kokkos_Bitset.hpp"
@@ -52,10 +54,8 @@
 /*********************************************************************************************************/
 // Include the architecture-specific kernel partial specializations here
 // NOTE: This needs to be outside all namespaces
+#include "TpetraExt_MatrixMatrix_Serial.hpp"
 #include "TpetraExt_MatrixMatrix_OpenMP.hpp"
-#include "TpetraExt_MatrixMatrix_Cuda.hpp"
-#include "TpetraExt_MatrixMatrix_HIP.hpp"
-#include "TpetraExt_MatrixMatrix_SYCL.hpp"
 
 namespace Tpetra {
 
@@ -1909,8 +1909,6 @@ void mult_A_B_newmatrix(BlockCrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal
   C = rcp(new block_crs_matrix_type(*graphC, values, Aview.blocksize));
 }
 
-/*********************************************************************************************************/
-// AB NewMatrix Kernel wrappers (Default non-threaded version for CrsMatrix)
 template <class Scalar,
           class LocalOrdinal,
           class GlobalOrdinal,
@@ -1918,200 +1916,201 @@ template <class Scalar,
           class LocalOrdinalViewType>
 void KernelWrappers<Scalar, LocalOrdinal, GlobalOrdinal, Node, LocalOrdinalViewType>::mult_A_B_newmatrix_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
                                                                                                                         CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
-                                                                                                                        const LocalOrdinalViewType& targetMapToOrigRow,
-                                                                                                                        const LocalOrdinalViewType& targetMapToImportRow,
+                                                                                                                        const LocalOrdinalViewType& Acol2Brow,
+                                                                                                                        const LocalOrdinalViewType& Acol2Irow,
                                                                                                                         const LocalOrdinalViewType& Bcol2Ccol,
                                                                                                                         const LocalOrdinalViewType& Icol2Ccol,
                                                                                                                         CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
                                                                                                                         Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node>> Cimport,
                                                                                                                         const std::string& label,
                                                                                                                         const Teuchos::RCP<Teuchos::ParameterList>& params) {
-  using Teuchos::Array;
-  using Teuchos::ArrayRCP;
-  using Teuchos::ArrayView;
   using Teuchos::RCP;
   using Teuchos::rcp;
+  RCP<Tpetra::Details::ProfilingRegion> MM = rcp(new Tpetra::Details::ProfilingRegion("TpetraExt: MMM: Newmatrix"));
 
-  Tpetra::Details::ProfilingRegion MM("TpetraExt: MMM: Newmatrix SerialCore");
+  // Node-specific code
+  std::string myalg("SPGEMM_KK_MEMORY");
+  int team_work_size;
+  std::string nodeLowerCase;
+  std::string nodeUpperCase;
+#ifdef KOKKOS_ENABLE_SERIAL
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosSerialWrapperNode>) {
+    nodeLowerCase  = "serial";
+    nodeUpperCase  = "Serial";
+    team_work_size = 1;
+    throw std::runtime_error("WE SHOULD NEVER GET HERE");
+  }
+#endif
+#ifdef KOKKOS_ENABLE_OPENMP
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosOpenMPWrapperNode>) {
+    nodeLowerCase  = "openmp";
+    nodeLowerCase  = "OpenMP";
+    team_work_size = 16;
+    throw std::runtime_error("WE SHOULD NEVER GET HERE");
+  }
+#endif
+#ifdef KOKKOS_ENABLE_CUDA
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosCudaWrapperNode>) {
+    nodeLowerCase  = "cuda";
+    nodeUpperCase  = "Cuda";
+    team_work_size = 16;
+  }
+#endif
+#ifdef KOKKOS_ENABLE_HIP
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosHipWrapperNode>) {
+    nodeLowerCase  = "hip";
+    nodeUpperCase  = "HIP";
+    team_work_size = 16;
+  }
+#endif
+#ifdef KOKKOS_ENABLE_SYCL
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosSYCLWrapperNode>) {
+    nodeLowerCase  = "sycl";
+    nodeUpperCase  = "SYCL";
+    team_work_size = 16;
+  }
+#endif
 
   // Lots and lots of typedefs
-  typedef typename Tpetra::CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>::local_matrix_host_type KCRS;
-  typedef typename KCRS::StaticCrsGraphType graph_t;
-  typedef typename graph_t::row_map_type::const_type c_lno_view_t;
-  typedef typename graph_t::row_map_type::non_const_type lno_view_t;
-  typedef typename graph_t::entries_type::non_const_type lno_nnz_view_t;
-  typedef typename KCRS::values_type::non_const_type scalar_view_t;
+  using KCRS = typename Tpetra::CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>::local_matrix_device_type;
+  using device_t = typename KCRS::device_type;
+  using graph_t = typename KCRS::StaticCrsGraphType;
+  using lno_view_t = typename graph_t::row_map_type::non_const_type;
+  using int_view_t = Kokkos::View<int*, typename lno_view_t::array_layout, typename lno_view_t::memory_space, typename lno_view_t::memory_traits>;
+  using c_lno_view_t = typename graph_t::row_map_type::const_type;
+  using lno_nnz_view_t = typename graph_t::entries_type::non_const_type;
+  using scalar_view_t = typename KCRS::values_type::non_const_type;
+  // typedef typename graph_t::row_map_type::const_type lno_view_t_const;
 
-  typedef Scalar SC;
-  typedef LocalOrdinal LO;
-  typedef GlobalOrdinal GO;
-  typedef Node NO;
-  typedef Map<LO, GO, NO> map_type;
-  const size_t ST_INVALID = Teuchos::OrdinalTraits<LO>::invalid();
-  const LO LO_INVALID     = Teuchos::OrdinalTraits<LO>::invalid();
-  const SC SC_ZERO        = Teuchos::ScalarTraits<Scalar>::zero();
-
-  // Sizes
-  RCP<const map_type> Ccolmap = C.getColMap();
-  size_t m                    = Aview.origMatrix->getLocalNumRows();
-  size_t n                    = Ccolmap->getLocalNumElements();
-  size_t b_max_nnz_per_row    = Bview.origMatrix->getLocalMaxNumRowEntries();
-
-  // Grab the  Kokkos::SparseCrsMatrices & inner stuff
-  const KCRS Amat = Aview.origMatrix->getLocalMatrixHost();
-  const KCRS Bmat = Bview.origMatrix->getLocalMatrixHost();
-
-  c_lno_view_t Arowptr = Amat.graph.row_map, Browptr = Bmat.graph.row_map;
-  const lno_nnz_view_t Acolind = Amat.graph.entries, Bcolind = Bmat.graph.entries;
-  const scalar_view_t Avals = Amat.values, Bvals = Bmat.values;
-
-  c_lno_view_t Irowptr;
-  lno_nnz_view_t Icolind;
-  scalar_view_t Ivals;
-  if (!Bview.importMatrix.is_null()) {
-    auto lclB         = Bview.importMatrix->getLocalMatrixHost();
-    Irowptr           = lclB.graph.row_map;
-    Icolind           = lclB.graph.entries;
-    Ivals             = lclB.values;
-    b_max_nnz_per_row = std::max(b_max_nnz_per_row, Bview.importMatrix->getLocalMaxNumRowEntries());
+  // Options
+  if (!params.is_null()) {
+    if (params->isParameter(nodeLowerCase + ": algorithm"))
+      myalg = params->get(nodeLowerCase + ": algorithm", myalg);
+    if (params->isParameter(nodeLowerCase + ": team work size"))
+      team_work_size = params->get(nodeLowerCase + ": team work size", team_work_size);
   }
 
-  // Classic csr assembly (low memory edition)
-  //
-  // mfh 27 Sep 2016: C_estimate_nnz does not promise an upper bound.
-  // The method loops over rows of A, and may resize after processing
-  // each row.  Chris Siefert says that this reflects experience in
-  // ML; for the non-threaded case, ML found it faster to spend less
-  // effort on estimation and risk an occasional reallocation.
-  size_t CSR_alloc = std::max(C_estimate_nnz(*Aview.origMatrix, *Bview.origMatrix), n);
-  lno_view_t Crowptr(Kokkos::ViewAllocateWithoutInitializing("Crowptr"), m + 1);
-  lno_nnz_view_t Ccolind(Kokkos::ViewAllocateWithoutInitializing("Ccolind"), CSR_alloc);
-  scalar_view_t Cvals(Kokkos::ViewAllocateWithoutInitializing("Cvals"), CSR_alloc);
+  // KokkosKernelsHandle
+  using KernelHandle = KokkosKernels::Experimental::KokkosKernelsHandle<
+      typename lno_view_t::const_value_type, typename lno_nnz_view_t::const_value_type, typename scalar_view_t::const_value_type,
+      typename device_t::execution_space, typename device_t::memory_space, typename device_t::memory_space>;
+  using IntKernelHandle = KokkosKernels::Experimental::KokkosKernelsHandle<
+      typename int_view_t::const_value_type, typename lno_nnz_view_t::const_value_type, typename scalar_view_t::const_value_type,
+      typename device_t::execution_space, typename device_t::memory_space, typename device_t::memory_space>;
 
-  // mfh 27 Sep 2016: The c_status array is an implementation detail
-  // of the local sparse matrix-matrix multiply routine.
+  // Grab the  Kokkos::SparseCrsMatrices
+  const KCRS Amat = Aview.origMatrix->getLocalMatrixDevice();
+  const KCRS Bmat = Bview.origMatrix->getLocalMatrixDevice();
 
-  // The status array will contain the index into colind where this entry was last deposited.
-  //   c_status[i] <  CSR_ip - not in the row yet
-  //   c_status[i] >= CSR_ip - this is the entry where you can find the data
-  // We start with this filled with INVALID's indicating that there are no entries yet.
-  // Sadly, this complicates the code due to the fact that size_t's are unsigned.
-  size_t INVALID = Teuchos::OrdinalTraits<size_t>::invalid();
-  std::vector<size_t> c_status(n, ST_INVALID);
+  c_lno_view_t Arowptr         = Amat.graph.row_map;
+  c_lno_view_t Browptr         = Bmat.graph.row_map;
+  const lno_nnz_view_t Acolind = Amat.graph.entries;
+  const lno_nnz_view_t Bcolind = Bmat.graph.entries;
+  const scalar_view_t Avals    = Amat.values;
+  const scalar_view_t Bvals    = Bmat.values;
 
-  // mfh 27 Sep 2016: Here is the local sparse matrix-matrix multiply
-  // routine.  The routine computes C := A * (B_local + B_remote).
-  //
-  // For column index Aik in row i of A, targetMapToOrigRow[Aik] tells
-  // you whether the corresponding row of B belongs to B_local
-  // ("orig") or B_remote ("Import").
+  // Get the algorithm mode
+  std::string alg = nodeUpperCase + std::string(" algorithm");
+  //  printf("DEBUG: Using kernel: %s\n",myalg.c_str());
+  if (!params.is_null() && params->isParameter(alg)) myalg = params->get(alg, myalg);
+  KokkosSparse::SPGEMMAlgorithm alg_enum = KokkosSparse::StringToSPGEMMAlgorithm(myalg);
 
-  // For each row of A/C
-  size_t CSR_ip = 0, OLD_ip = 0;
-  for (size_t i = 0; i < m; i++) {
-    // mfh 27 Sep 2016: m is the number of rows in the input matrix A
-    // on the calling process.
-    Crowptr[i] = CSR_ip;
+  // Merge the B and Bimport matrices
+  KCRS Bmerged = Tpetra::MMdetails::merge_matrices(Aview, Bview, Acol2Brow, Acol2Irow, Bcol2Ccol, Icol2Ccol, C.getColMap()->getLocalNumElements());
 
-    // mfh 27 Sep 2016: For each entry of A in the current row of A
-    for (size_t k = Arowptr[i]; k < Arowptr[i + 1]; k++) {
-      LO Aik        = Acolind[k];  // local column index of current entry of A
-      const SC Aval = Avals[k];    // value of current entry of A
-      if (Aval == SC_ZERO)
-        continue;  // skip explicitly stored zero values in A
+  if constexpr (Details::MatrixTraits<Scalar, LocalOrdinal, GlobalOrdinal, Node>::spgemmNeedsSortedInputs()) {
+    if (!KokkosSparse::isCrsGraphSorted(Bmerged.graph.row_map, Bmerged.graph.entries)) {
+      KokkosSparse::sort_crs_matrix(Bmerged);
+    }
+  }
 
-      if (targetMapToOrigRow[Aik] != LO_INVALID) {
-        // mfh 27 Sep 2016: If the entry of targetMapToOrigRow
-        // corresponding to the current entry of A is populated, then
-        // the corresponding row of B is in B_local (i.e., it lives on
-        // the calling process).
+  MM = Teuchos::null;
+  MM = rcp(new Tpetra::Details::ProfilingRegion("TpetraExt: MMM: Newmatrix Core"));
 
-        // Local matrix
-        size_t Bk = static_cast<size_t>(targetMapToOrigRow[Aik]);
+  // Do the multiply on whatever we've got
+  typename KernelHandle::nnz_lno_t AnumRows = Amat.numRows();
+  typename KernelHandle::nnz_lno_t BnumRows = Bmerged.numRows();
+  typename KernelHandle::nnz_lno_t BnumCols = Bmerged.numCols();
 
-        // mfh 27 Sep 2016: Go through all entries in that row of B_local.
-        for (size_t j = Browptr[Bk]; j < Browptr[Bk + 1]; ++j) {
-          LO Bkj = Bcolind[j];
-          LO Cij = Bcol2Ccol[Bkj];
+  // regardless of whether integer row ptrs are used, need to ultimately produce the row pointer, entries, and values of the expected types
+  lno_view_t row_mapC(Kokkos::ViewAllocateWithoutInitializing("non_const_lno_row"), AnumRows + 1);
+  lno_nnz_view_t entriesC;
+  scalar_view_t valuesC;
 
-          if (c_status[Cij] == INVALID || c_status[Cij] < OLD_ip) {
-            // New entry
-            c_status[Cij]   = CSR_ip;
-            Ccolind[CSR_ip] = Cij;
-            Cvals[CSR_ip]   = Aval * Bvals[j];
-            CSR_ip++;
+  Tpetra::Details::IntRowPtrHelper<decltype(Bmerged)> irph(Bmerged.nnz(), Bmerged.graph.row_map);
+  const bool useIntRowptrs =
+      irph.shouldUseIntRowptrs() &&
+      Aview.origMatrix->getApplyHelper()->shouldUseIntRowptrs();
 
-          } else {
-            Cvals[c_status[Cij]] += Aval * Bvals[j];
-          }
-        }
+  if (useIntRowptrs) {
+    IntKernelHandle kh;
+    kh.create_spgemm_handle(alg_enum);
+    kh.set_team_work_size(team_work_size);
 
-      } else {
-        // mfh 27 Sep 2016: If the entry of targetMapToOrigRow
-        // corresponding to the current entry of A NOT populated (has
-        // a flag "invalid" value), then the corresponding row of B is
-        // in B_local (i.e., it lives on the calling process).
+    int_view_t int_row_mapC(Kokkos::ViewAllocateWithoutInitializing("non_const_int_row"), AnumRows + 1);
 
-        // Remote matrix
-        size_t Ik = static_cast<size_t>(targetMapToImportRow[Aik]);
-        for (size_t j = Irowptr[Ik]; j < Irowptr[Ik + 1]; ++j) {
-          LO Ikj = Icolind[j];
-          LO Cij = Icol2Ccol[Ikj];
+    auto Aint = Aview.origMatrix->getApplyHelper()->getIntRowptrMatrix(Amat);
+    auto Bint = irph.getIntRowptrMatrix(Bmerged);
 
-          if (c_status[Cij] == INVALID || c_status[Cij] < OLD_ip) {
-            // New entry
-            c_status[Cij]   = CSR_ip;
-            Ccolind[CSR_ip] = Cij;
-            Cvals[CSR_ip]   = Aval * Ivals[j];
-            CSR_ip++;
-          } else {
-            Cvals[c_status[Cij]] += Aval * Ivals[j];
-          }
-        }
-      }
+    {
+      Tpetra::Details::ProfilingRegion MM2("TpetraExt: MMM: Newmatrix KokkosKernels symbolic int");
+      KokkosSparse::spgemm_symbolic(&kh, AnumRows, BnumRows, BnumCols, Aint.graph.row_map, Aint.graph.entries, false, Bint.graph.row_map, Bint.graph.entries, false, int_row_mapC);
     }
 
-    // Resize for next pass if needed
-    if (i + 1 < m && CSR_ip + std::min(n, (Arowptr[i + 2] - Arowptr[i + 1]) * b_max_nnz_per_row) > CSR_alloc) {
-      CSR_alloc *= 2;
-      Kokkos::resize(Ccolind, CSR_alloc);
-      Kokkos::resize(Cvals, CSR_alloc);
+    Tpetra::Details::ProfilingRegion MM2("TpetraExt: MMM: Newmatrix KokkosKernels numeric int");
+
+    size_t c_nnz_size = kh.get_spgemm_handle()->get_c_nnz();
+    if (c_nnz_size) {
+      entriesC = lno_nnz_view_t(Kokkos::ViewAllocateWithoutInitializing("entriesC"), c_nnz_size);
+      valuesC  = scalar_view_t(Kokkos::ViewAllocateWithoutInitializing("valuesC"), c_nnz_size);
     }
-    OLD_ip = CSR_ip;
+    KokkosSparse::spgemm_numeric(&kh, AnumRows, BnumRows, BnumCols, Aint.graph.row_map, Aint.graph.entries, Aint.values, false, Bint.graph.row_map, Bint.graph.entries, Bint.values, false, int_row_mapC, entriesC, valuesC);
+    // transfer the integer rowptrs back to the correct rowptr type
+    Kokkos::parallel_for(
+        int_row_mapC.size(), KOKKOS_LAMBDA(int i) { row_mapC(i) = int_row_mapC(i); });
+    kh.destroy_spgemm_handle();
+
+  } else {
+    KernelHandle kh;
+    kh.create_spgemm_handle(alg_enum);
+    kh.set_team_work_size(team_work_size);
+
+    {
+      Tpetra::Details::ProfilingRegion MM2("TpetraExt: Jacobi: Newmatrix KokkosKernels symbolic non-int");
+      KokkosSparse::spgemm_symbolic(&kh, AnumRows, BnumRows, BnumCols, Amat.graph.row_map, Amat.graph.entries, false, Bmerged.graph.row_map, Bmerged.graph.entries, false, row_mapC);
+    }
+
+    Tpetra::Details::ProfilingRegion MM2("TpetraExt: Jacobi: Newmatrix KokkosKernels numeric non-int");
+    size_t c_nnz_size = kh.get_spgemm_handle()->get_c_nnz();
+    if (c_nnz_size) {
+      entriesC = lno_nnz_view_t(Kokkos::ViewAllocateWithoutInitializing("entriesC"), c_nnz_size);
+      valuesC  = scalar_view_t(Kokkos::ViewAllocateWithoutInitializing("valuesC"), c_nnz_size);
+    }
+
+    KokkosSparse::spgemm_numeric(&kh, AnumRows, BnumRows, BnumCols, Amat.graph.row_map, Amat.graph.entries, Amat.values, false, Bmerged.graph.row_map, Bmerged.graph.entries, Bmerged.values, false, row_mapC, entriesC, valuesC);
+
+    kh.destroy_spgemm_handle();
   }
 
-  Crowptr[m] = CSR_ip;
+  MM = Teuchos::null;
+  MM = rcp(new Tpetra::Details::ProfilingRegion("TpetraExt: MMM: Newmatrix Sort"));
 
-  // Downward resize
-  Kokkos::resize(Ccolind, CSR_ip);
-  Kokkos::resize(Cvals, CSR_ip);
+  // Sort & set values
+  if (params.is_null() || params->get("sort entries", true))
+    Import_Util::sortCrsEntries(row_mapC, entriesC, valuesC);
+  C.setAllValues(row_mapC, entriesC, valuesC);
 
-  {
-    Tpetra::Details::ProfilingRegion MM3("TpetraExt: MMM: Newmatrix Final Sort");
+  MM = Teuchos::null;
+  MM = rcp(new Tpetra::Details::ProfilingRegion("TpetraExt: MMM: Newmatrix ESFC"));
 
-    // Final sort & set of CRS arrays
-    if (params.is_null() || params->get("sort entries", Details::MatrixTraits<Scalar, LocalOrdinal, GlobalOrdinal, Node>::spgemmNeedsSortedInputs()))
-      Import_Util::sortCrsEntries(Crowptr, Ccolind, Cvals);
-    C.setAllValues(Crowptr, Ccolind, Cvals);
-  }
-
-  Tpetra::Details::ProfilingRegion MM4("TpetraExt: MMM: Newmatrix ESCC");
-  {
-    // Final FillComplete
-    //
-    // mfh 27 Sep 2016: So-called "expert static fill complete" bypasses
-    // Import (from domain Map to column Map) construction (which costs
-    // lots of communication) by taking the previously constructed
-    // Import object.  We should be able to do this without interfering
-    // with the implementation of the local part of sparse matrix-matrix
-    // multply above.
-    RCP<Teuchos::ParameterList> labelList = rcp(new Teuchos::ParameterList);
-    labelList->set("Timer Label", label);
-    if (!params.is_null()) labelList->set("compute global constants", params->get("compute global constants", true));
-    RCP<const Export<LO, GO, NO>> dummyExport;
-    C.expertStaticFillComplete(Bview.origMatrix->getDomainMap(), Aview.origMatrix->getRangeMap(), Cimport, dummyExport, labelList);
-  }
+  // Final Fillcomplete
+  RCP<Teuchos::ParameterList> labelList = rcp(new Teuchos::ParameterList);
+  labelList->set("Timer Label", label);
+  if (!params.is_null()) labelList->set("compute global constants", params->get("compute global constants", true));
+  RCP<const Export<LocalOrdinal, GlobalOrdinal, Node>> dummyExport;
+  C.expertStaticFillComplete(Bview.origMatrix->getDomainMap(), Aview.origMatrix->getRangeMap(), Cimport, dummyExport, labelList);
 }
+
 /*********************************************************************************************************/
 // Kernel method for computing the local portion of C = A*B (reuse)
 template <class Scalar,
@@ -2215,128 +2214,15 @@ template <class Scalar,
           class LocalOrdinalViewType>
 void KernelWrappers<Scalar, LocalOrdinal, GlobalOrdinal, Node, LocalOrdinalViewType>::mult_A_B_reuse_kernel_wrapper(CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
                                                                                                                     CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
-                                                                                                                    const LocalOrdinalViewType& targetMapToOrigRow,
-                                                                                                                    const LocalOrdinalViewType& targetMapToImportRow,
-                                                                                                                    const LocalOrdinalViewType& Bcol2Ccol,
-                                                                                                                    const LocalOrdinalViewType& Icol2Ccol,
+                                                                                                                    const LocalOrdinalViewType& targetMapToOrigRow_dev,
+                                                                                                                    const LocalOrdinalViewType& targetMapToImportRow_dev,
+                                                                                                                    const LocalOrdinalViewType& Bcol2Ccol_dev,
+                                                                                                                    const LocalOrdinalViewType& Icol2Ccol_dev,
                                                                                                                     CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
-                                                                                                                    Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node>> /* Cimport */,
+                                                                                                                    Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node>> Cimport,
                                                                                                                     const std::string& label,
-                                                                                                                    const Teuchos::RCP<Teuchos::ParameterList>& /* params */) {
-  using Teuchos::RCP;
-  using Teuchos::rcp;
-
-  // Lots and lots of typedefs
-  typedef typename Tpetra::CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>::local_matrix_host_type KCRS;
-  typedef typename KCRS::StaticCrsGraphType graph_t;
-  typedef typename graph_t::row_map_type::const_type c_lno_view_t;
-  typedef typename graph_t::entries_type::non_const_type lno_nnz_view_t;
-  typedef typename KCRS::values_type::non_const_type scalar_view_t;
-
-  typedef Scalar SC;
-  typedef LocalOrdinal LO;
-  typedef GlobalOrdinal GO;
-  typedef Node NO;
-  typedef Map<LO, GO, NO> map_type;
-  const size_t ST_INVALID = Teuchos::OrdinalTraits<LO>::invalid();
-  const LO LO_INVALID     = Teuchos::OrdinalTraits<LO>::invalid();
-  const SC SC_ZERO        = Teuchos::ScalarTraits<Scalar>::zero();
-
-  Tpetra::Details::ProfilingRegion MM("TpetraExt: MMM: Reuse SerialCore");
-  (void)label;
-
-  // Sizes
-  RCP<const map_type> Ccolmap = C.getColMap();
-  size_t m                    = Aview.origMatrix->getLocalNumRows();
-  size_t n                    = Ccolmap->getLocalNumElements();
-
-  // Grab the  Kokkos::SparseCrsMatrices & inner stuff
-  const KCRS Amat = Aview.origMatrix->getLocalMatrixHost();
-  const KCRS Bmat = Bview.origMatrix->getLocalMatrixHost();
-  const KCRS Cmat = C.getLocalMatrixHost();
-
-  c_lno_view_t Arowptr = Amat.graph.row_map, Browptr = Bmat.graph.row_map, Crowptr = Cmat.graph.row_map;
-  const lno_nnz_view_t Acolind = Amat.graph.entries, Bcolind = Bmat.graph.entries, Ccolind = Cmat.graph.entries;
-  const scalar_view_t Avals = Amat.values, Bvals = Bmat.values;
-  scalar_view_t Cvals = Cmat.values;
-
-  c_lno_view_t Irowptr;
-  lno_nnz_view_t Icolind;
-  scalar_view_t Ivals;
-  if (!Bview.importMatrix.is_null()) {
-    auto lclB = Bview.importMatrix->getLocalMatrixHost();
-    Irowptr   = lclB.graph.row_map;
-    Icolind   = lclB.graph.entries;
-    Ivals     = lclB.values;
-  }
-
-  // Classic csr assembly (low memory edition)
-  // mfh 27 Sep 2016: The c_status array is an implementation detail
-  // of the local sparse matrix-matrix multiply routine.
-
-  // The status array will contain the index into colind where this entry was last deposited.
-  //   c_status[i] <  CSR_ip - not in the row yet
-  //   c_status[i] >= CSR_ip - this is the entry where you can find the data
-  // We start with this filled with INVALID's indicating that there are no entries yet.
-  // Sadly, this complicates the code due to the fact that size_t's are unsigned.
-  std::vector<size_t> c_status(n, ST_INVALID);
-
-  // For each row of A/C
-  size_t CSR_ip = 0, OLD_ip = 0;
-  for (size_t i = 0; i < m; i++) {
-    // First fill the c_status array w/ locations where we're allowed to
-    // generate nonzeros for this row
-    OLD_ip = Crowptr[i];
-    CSR_ip = Crowptr[i + 1];
-    for (size_t k = OLD_ip; k < CSR_ip; k++) {
-      c_status[Ccolind[k]] = k;
-
-      // Reset values in the row of C
-      Cvals[k] = SC_ZERO;
-    }
-
-    for (size_t k = Arowptr[i]; k < Arowptr[i + 1]; k++) {
-      LO Aik        = Acolind[k];
-      const SC Aval = Avals[k];
-      if (Aval == SC_ZERO)
-        continue;
-
-      if (targetMapToOrigRow[Aik] != LO_INVALID) {
-        // Local matrix
-        size_t Bk = static_cast<size_t>(targetMapToOrigRow[Aik]);
-
-        for (size_t j = Browptr[Bk]; j < Browptr[Bk + 1]; ++j) {
-          LO Bkj = Bcolind[j];
-          LO Cij = Bcol2Ccol[Bkj];
-
-          TEUCHOS_TEST_FOR_EXCEPTION(c_status[Cij] < OLD_ip || c_status[Cij] >= CSR_ip,
-                                     std::runtime_error, "Trying to insert a new entry (" << i << "," << Cij << ") into a static graph "
-                                                                                          << "(c_status = " << c_status[Cij] << " of [" << OLD_ip << "," << CSR_ip << "))");
-
-          Cvals[c_status[Cij]] += Aval * Bvals[j];
-        }
-
-      } else {
-        // Remote matrix
-        size_t Ik = static_cast<size_t>(targetMapToImportRow[Aik]);
-        for (size_t j = Irowptr[Ik]; j < Irowptr[Ik + 1]; ++j) {
-          LO Ikj = Icolind[j];
-          LO Cij = Icol2Ccol[Ikj];
-
-          TEUCHOS_TEST_FOR_EXCEPTION(c_status[Cij] < OLD_ip || c_status[Cij] >= CSR_ip,
-                                     std::runtime_error, "Trying to insert a new entry (" << i << "," << Cij << ") into a static graph "
-                                                                                          << "(c_status = " << c_status[Cij] << " of [" << OLD_ip << "," << CSR_ip << "))");
-
-          Cvals[c_status[Cij]] += Aval * Ivals[j];
-        }
-      }
-    }
-  }
-
-  {
-    Tpetra::Details::ProfilingRegion MM3("TpetraExt: MMM: Reuse ESFC");
-    C.fillComplete(C.getDomainMap(), C.getRangeMap());
-  }
+                                                                                                                    const Teuchos::RCP<Teuchos::ParameterList>& params) {
+  mult_A_B_reuse_kernel_wrapper_fun(Aview, Bview, targetMapToOrigRow_dev, targetMapToImportRow_dev, Bcol2Ccol_dev, Icol2Ccol_dev, C, Cimport, label, params);
 }
 
 /*********************************************************************************************************/
@@ -2499,9 +2385,6 @@ void jacobi_A_B_newmatrix(
 }
 
 /*********************************************************************************************************/
-// Jacobi AB NewMatrix Kernel wrappers (Default non-threaded version)
-// Kernel method for computing the local portion of C = (I-omega D^{-1} A)*B
-
 template <class Scalar,
           class LocalOrdinal,
           class GlobalOrdinal,
@@ -2511,222 +2394,86 @@ void KernelWrappers2<Scalar, LocalOrdinal, GlobalOrdinal, Node, LocalOrdinalView
                                                                                                                            const Vector<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Dinv,
                                                                                                                            CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
                                                                                                                            CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
-                                                                                                                           const LocalOrdinalViewType& targetMapToOrigRow,
-                                                                                                                           const LocalOrdinalViewType& targetMapToImportRow,
+                                                                                                                           const LocalOrdinalViewType& Acol2Brow,
+                                                                                                                           const LocalOrdinalViewType& Acol2Irow,
                                                                                                                            const LocalOrdinalViewType& Bcol2Ccol,
                                                                                                                            const LocalOrdinalViewType& Icol2Ccol,
                                                                                                                            CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
                                                                                                                            Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node>> Cimport,
                                                                                                                            const std::string& label,
                                                                                                                            const Teuchos::RCP<Teuchos::ParameterList>& params) {
-  Tpetra::Details::ProfilingRegion MM("TpetraExt: Jacobi: Newmatrix SerialCore");
+  // Node-specific code
+  std::string myalg("KK");
+  int team_work_size;
+  std::string nodeLowerCase;
+  std::string nodeUpperCase;
+#ifdef KOKKOS_ENABLE_SERIAL
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosSerialWrapperNode>) {
+    nodeLowerCase  = "serial";
+    nodeUpperCase  = "Serial";
+    team_work_size = 1;
+    throw std::runtime_error("WE SHOULD NEVER GET HERE");
+  }
+#endif
+#ifdef KOKKOS_ENABLE_OPENMP
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosOpenMPWrapperNode>) {
+    nodeLowerCase  = "openmp";
+    nodeLowerCase  = "OpenMP";
+    team_work_size = 16;
+    throw std::runtime_error("WE SHOULD NEVER GET HERE");
+  }
+#endif
+#ifdef KOKKOS_ENABLE_CUDA
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosCudaWrapperNode>) {
+    nodeLowerCase  = "cuda";
+    nodeUpperCase  = "Cuda";
+    team_work_size = 16;
+  }
+#endif
+#ifdef KOKKOS_ENABLE_HIP
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosHipWrapperNode>) {
+    nodeLowerCase  = "hip";
+    nodeUpperCase  = "HIP";
+    team_work_size = 16;
+  }
+#endif
+#ifdef KOKKOS_ENABLE_SYCL
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosSYCLWrapperNode>) {
+    nodeLowerCase  = "sycl";
+    nodeUpperCase  = "SYCL";
+    team_work_size = 16;
+  }
+#endif
 
-  using Teuchos::Array;
-  using Teuchos::ArrayRCP;
-  using Teuchos::ArrayView;
   using Teuchos::RCP;
   using Teuchos::rcp;
+  RCP<Tpetra::Details::ProfilingRegion> MM = rcp(new Tpetra::Details::ProfilingRegion("TpetraExt: MMM: Jacobi"));
 
-  // Lots and lots of typedefs
-  typedef typename Tpetra::CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>::local_matrix_host_type KCRS;
-  typedef typename KCRS::StaticCrsGraphType graph_t;
-  typedef typename graph_t::row_map_type::const_type c_lno_view_t;
-  typedef typename graph_t::row_map_type::non_const_type lno_view_t;
-  typedef typename graph_t::entries_type::non_const_type lno_nnz_view_t;
-  typedef typename KCRS::values_type::non_const_type scalar_view_t;
-
-  // Jacobi-specific
-  typedef typename scalar_view_t::memory_space scalar_memory_space;
-
-  typedef Scalar SC;
-  typedef LocalOrdinal LO;
-  typedef GlobalOrdinal GO;
-  typedef Node NO;
-
-  typedef Map<LO, GO, NO> map_type;
-  size_t ST_INVALID = Teuchos::OrdinalTraits<LO>::invalid();
-  LO LO_INVALID     = Teuchos::OrdinalTraits<LO>::invalid();
-
-  // Sizes
-  RCP<const map_type> Ccolmap = C.getColMap();
-  size_t m                    = Aview.origMatrix->getLocalNumRows();
-  size_t n                    = Ccolmap->getLocalNumElements();
-  size_t b_max_nnz_per_row    = Bview.origMatrix->getLocalMaxNumRowEntries();
-
-  // Grab the  Kokkos::SparseCrsMatrices & inner stuff
-  const KCRS Amat = Aview.origMatrix->getLocalMatrixHost();
-  const KCRS Bmat = Bview.origMatrix->getLocalMatrixHost();
-
-  c_lno_view_t Arowptr = Amat.graph.row_map, Browptr = Bmat.graph.row_map;
-  const lno_nnz_view_t Acolind = Amat.graph.entries, Bcolind = Bmat.graph.entries;
-  const scalar_view_t Avals = Amat.values, Bvals = Bmat.values;
-
-  c_lno_view_t Irowptr;
-  lno_nnz_view_t Icolind;
-  scalar_view_t Ivals;
-  if (!Bview.importMatrix.is_null()) {
-    auto lclB         = Bview.importMatrix->getLocalMatrixHost();
-    Irowptr           = lclB.graph.row_map;
-    Icolind           = lclB.graph.entries;
-    Ivals             = lclB.values;
-    b_max_nnz_per_row = std::max(b_max_nnz_per_row, Bview.importMatrix->getLocalMaxNumRowEntries());
+  // Options
+  if (!params.is_null()) {
+    if (params->isParameter(nodeLowerCase + ": jacobi algorithm"))
+      myalg = params->get(nodeLowerCase + ": jacobi algorithm", myalg);
   }
 
-  // Jacobi-specific inner stuff
-  auto Dvals =
-      Dinv.template getLocalView<scalar_memory_space>(Access::ReadOnly);
-
-  // Teuchos::ArrayView::operator[].
-  // The status array will contain the index into colind where this entry was last deposited.
-  // c_status[i] < CSR_ip - not in the row yet.
-  // c_status[i] >= CSR_ip, this is the entry where you can find the data
-  // We start with this filled with INVALID's indicating that there are no entries yet.
-  // Sadly, this complicates the code due to the fact that size_t's are unsigned.
-  size_t INVALID = Teuchos::OrdinalTraits<size_t>::invalid();
-  Array<size_t> c_status(n, ST_INVALID);
-
-  // Classic csr assembly (low memory edition)
-  //
-  // mfh 27 Sep 2016: C_estimate_nnz does not promise an upper bound.
-  // The method loops over rows of A, and may resize after processing
-  // each row.  Chris Siefert says that this reflects experience in
-  // ML; for the non-threaded case, ML found it faster to spend less
-  // effort on estimation and risk an occasional reallocation.
-  size_t CSR_alloc = std::max(C_estimate_nnz(*Aview.origMatrix, *Bview.origMatrix), n);
-  lno_view_t Crowptr(Kokkos::ViewAllocateWithoutInitializing("Crowptr"), m + 1);
-  lno_nnz_view_t Ccolind(Kokkos::ViewAllocateWithoutInitializing("Ccolind"), CSR_alloc);
-  scalar_view_t Cvals(Kokkos::ViewAllocateWithoutInitializing("Cvals"), CSR_alloc);
-  size_t CSR_ip = 0, OLD_ip = 0;
-
-  const SC SC_ZERO = Teuchos::ScalarTraits<Scalar>::zero();
-
-  // mfh 27 Sep 2016: Here is the local sparse matrix-matrix multiply
-  // routine.  The routine computes
-  //
-  // C := (I - omega * D^{-1} * A) * (B_local + B_remote)).
-  //
-  // This corresponds to one sweep of (weighted) Jacobi.
-  //
-  // For column index Aik in row i of A, targetMapToOrigRow[Aik] tells
-  // you whether the corresponding row of B belongs to B_local
-  // ("orig") or B_remote ("Import").
-
-  // For each row of A/C
-  for (size_t i = 0; i < m; i++) {
-    // mfh 27 Sep 2016: m is the number of rows in the input matrix A
-    // on the calling process.
-    Crowptr[i]        = CSR_ip;
-    SC minusOmegaDval = -omega * Dvals(i, 0);
-
-    // Entries of B
-    for (size_t j = Browptr[i]; j < Browptr[i + 1]; j++) {
-      Scalar Bval = Bvals[j];
-      if (Bval == SC_ZERO)
-        continue;
-      LO Bij = Bcolind[j];
-      LO Cij = Bcol2Ccol[Bij];
-
-      // Assume no repeated entries in B
-      c_status[Cij]   = CSR_ip;
-      Ccolind[CSR_ip] = Cij;
-      Cvals[CSR_ip]   = Bvals[j];
-      CSR_ip++;
-    }
-
-    // Entries of -omega * Dinv * A * B
-    for (size_t k = Arowptr[i]; k < Arowptr[i + 1]; k++) {
-      LO Aik        = Acolind[k];
-      const SC Aval = Avals[k];
-      if (Aval == SC_ZERO)
-        continue;
-
-      if (targetMapToOrigRow[Aik] != LO_INVALID) {
-        // Local matrix
-        size_t Bk = static_cast<size_t>(targetMapToOrigRow[Aik]);
-
-        for (size_t j = Browptr[Bk]; j < Browptr[Bk + 1]; ++j) {
-          LO Bkj = Bcolind[j];
-          LO Cij = Bcol2Ccol[Bkj];
-
-          if (c_status[Cij] == INVALID || c_status[Cij] < OLD_ip) {
-            // New entry
-            c_status[Cij]   = CSR_ip;
-            Ccolind[CSR_ip] = Cij;
-            Cvals[CSR_ip]   = minusOmegaDval * Aval * Bvals[j];
-            CSR_ip++;
-
-          } else {
-            Cvals[c_status[Cij]] += minusOmegaDval * Aval * Bvals[j];
-          }
-        }
-
-      } else {
-        // Remote matrix
-        size_t Ik = static_cast<size_t>(targetMapToImportRow[Aik]);
-        for (size_t j = Irowptr[Ik]; j < Irowptr[Ik + 1]; ++j) {
-          LO Ikj = Icolind[j];
-          LO Cij = Icol2Ccol[Ikj];
-
-          if (c_status[Cij] == INVALID || c_status[Cij] < OLD_ip) {
-            // New entry
-            c_status[Cij]   = CSR_ip;
-            Ccolind[CSR_ip] = Cij;
-            Cvals[CSR_ip]   = minusOmegaDval * Aval * Ivals[j];
-            CSR_ip++;
-          } else {
-            Cvals[c_status[Cij]] += minusOmegaDval * Aval * Ivals[j];
-          }
-        }
-      }
-    }
-
-    // Resize for next pass if needed
-    if (i + 1 < m && CSR_ip + std::min(n, (Arowptr[i + 2] - Arowptr[i + 1] + 1) * b_max_nnz_per_row) > CSR_alloc) {
-      CSR_alloc *= 2;
-      Kokkos::resize(Ccolind, CSR_alloc);
-      Kokkos::resize(Cvals, CSR_alloc);
-    }
-    OLD_ip = CSR_ip;
+  if (myalg == "MSAK") {
+    ::Tpetra::MatrixMatrix::ExtraKernels::jacobi_A_B_newmatrix_MultiplyScaleAddKernel(omega, Dinv, Aview, Bview, Acol2Brow, Acol2Irow, Bcol2Ccol, Icol2Ccol, C, Cimport, label, params);
+  } else if (myalg == "KK") {
+    jacobi_A_B_newmatrix_KokkosKernels(omega, Dinv, Aview, Bview, Acol2Brow, Acol2Irow, Bcol2Ccol, Icol2Ccol, C, Cimport, label, params);
+  } else {
+    throw std::runtime_error("Tpetra::MatrixMatrix::Jacobi newmatrix unknown kernel");
   }
-  Crowptr[m] = CSR_ip;
 
-  // Downward resize
-  Kokkos::resize(Ccolind, CSR_ip);
-  Kokkos::resize(Cvals, CSR_ip);
+  MM = Teuchos::null;
+  MM = rcp(new Tpetra::Details::ProfilingRegion("TpetraExt: Jacobi: Newmatrix ESFC"));
 
-  {
-    Tpetra::Details::ProfilingRegion MM2("TpetraExt: Jacobi: Newmatrix Final Sort");
+  // Final Fillcomplete
+  RCP<Teuchos::ParameterList> labelList = rcp(new Teuchos::ParameterList);
+  labelList->set("Timer Label", label);
+  if (!params.is_null()) labelList->set("compute global constants", params->get("compute global constants", true));
 
-    // Replace the column map
-    //
-    // mfh 27 Sep 2016: We do this because C was originally created
-    // without a column Map.  Now we have its column Map.
-    C.replaceColMap(Ccolmap);
-
-    // Final sort & set of CRS arrays
-    //
-    // TODO (mfh 27 Sep 2016) Will the thread-parallel "local" sparse
-    // matrix-matrix multiply routine sort the entries for us?
-    // Final sort & set of CRS arrays
-    if (params.is_null() || params->get("sort entries", Details::MatrixTraits<Scalar, LocalOrdinal, GlobalOrdinal, Node>::spgemmNeedsSortedInputs()))
-      Import_Util::sortCrsEntries(Crowptr, Ccolind, Cvals);
-    C.setAllValues(Crowptr, Ccolind, Cvals);
-  }
-  {
-    Tpetra::Details::ProfilingRegion MM3("TpetraExt: Jacobi: Newmatrix ESFC");
-
-    // Final FillComplete
-    //
-    // mfh 27 Sep 2016: So-called "expert static fill complete" bypasses
-    // Import (from domain Map to column Map) construction (which costs
-    // lots of communication) by taking the previously constructed
-    // Import object.  We should be able to do this without interfering
-    // with the implementation of the local part of sparse matrix-matrix
-    // multply above
-    RCP<Teuchos::ParameterList> labelList = rcp(new Teuchos::ParameterList);
-    labelList->set("Timer Label", label);
-    if (!params.is_null()) labelList->set("compute global constants", params->get("compute global constants", true));
-    RCP<const Export<LO, GO, NO>> dummyExport;
+  // NOTE: MSAK already fillCompletes, so we have to check here
+  if (!C.isFillComplete()) {
+    RCP<const Export<LocalOrdinal, GlobalOrdinal, Node>> dummyExport;
     C.expertStaticFillComplete(Bview.origMatrix->getDomainMap(), Aview.origMatrix->getRangeMap(), Cimport, dummyExport, labelList);
   }
 }
@@ -2842,144 +2589,235 @@ void KernelWrappers2<Scalar, LocalOrdinal, GlobalOrdinal, Node, LocalOrdinalView
                                                                                                                        const Vector<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Dinv,
                                                                                                                        CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
                                                                                                                        CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
-                                                                                                                       const LocalOrdinalViewType& targetMapToOrigRow,
-                                                                                                                       const LocalOrdinalViewType& targetMapToImportRow,
-                                                                                                                       const LocalOrdinalViewType& Bcol2Ccol,
-                                                                                                                       const LocalOrdinalViewType& Icol2Ccol,
+                                                                                                                       const LocalOrdinalViewType& targetMapToOrigRow_dev,
+                                                                                                                       const LocalOrdinalViewType& targetMapToImportRow_dev,
+                                                                                                                       const LocalOrdinalViewType& Bcol2Ccol_dev,
+                                                                                                                       const LocalOrdinalViewType& Icol2Ccol_dev,
                                                                                                                        CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
-                                                                                                                       Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node>> /* Cimport */,
+                                                                                                                       Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node>> Cimport,
                                                                                                                        const std::string& label,
-                                                                                                                       const Teuchos::RCP<Teuchos::ParameterList>& /* params */) {
-  Tpetra::Details::ProfilingRegion MM2("TpetraExt: Jacobi: Reuse Serial Core");
-  (void)label;
+                                                                                                                       const Teuchos::RCP<Teuchos::ParameterList>& params) {
+  jacobi_A_B_reuse_kernel_wrapper_fun(omega, Dinv, Aview, Bview, targetMapToOrigRow_dev, targetMapToImportRow_dev, Bcol2Ccol_dev, Icol2Ccol_dev, C, Cimport, label, params);
+}
+
+/*********************************************************************************************************/
+template <class Scalar,
+          class LocalOrdinal,
+          class GlobalOrdinal,
+          class Node,
+          class LocalOrdinalViewType>
+void KernelWrappers2<Scalar, LocalOrdinal, GlobalOrdinal, Node, LocalOrdinalViewType>::jacobi_A_B_newmatrix_KokkosKernels(Scalar omega,
+                                                                                                                          const Vector<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Dinv,
+                                                                                                                          CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Aview,
+                                                                                                                          CrsMatrixStruct<Scalar, LocalOrdinal, GlobalOrdinal, Node>& Bview,
+                                                                                                                          const LocalOrdinalViewType& Acol2Brow,
+                                                                                                                          const LocalOrdinalViewType& Acol2Irow,
+                                                                                                                          const LocalOrdinalViewType& Bcol2Ccol,
+                                                                                                                          const LocalOrdinalViewType& Icol2Ccol,
+                                                                                                                          CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& C,
+                                                                                                                          Teuchos::RCP<const Import<LocalOrdinal, GlobalOrdinal, Node>> Cimport,
+                                                                                                                          const std::string& label,
+                                                                                                                          const Teuchos::RCP<Teuchos::ParameterList>& params) {
+  // Check if the diagonal entries exist in debug mode
+  const bool debug = Tpetra::Details::Behavior::debug();
+  if (debug) {
+    auto rowMap = Aview.origMatrix->getRowMap();
+    Tpetra::Vector<Scalar, LocalOrdinal, GlobalOrdinal, Node> diags(rowMap);
+    Aview.origMatrix->getLocalDiagCopy(diags);
+    size_t diagLength = rowMap->getLocalNumElements();
+    Teuchos::Array<Scalar> diagonal(diagLength);
+    diags.get1dCopy(diagonal());
+
+    for (size_t i = 0; i < diagLength; ++i) {
+      TEUCHOS_TEST_FOR_EXCEPTION(diagonal[i] == Teuchos::ScalarTraits<Scalar>::zero(),
+                                 std::runtime_error,
+                                 "Matrix A has a zero/missing diagonal: " << diagonal[i] << std::endl
+                                                                          << "KokkosKernels Jacobi-fused SpGEMM requires nonzero diagonal entries in A" << std::endl);
+    }
+  }
 
   using Teuchos::RCP;
   using Teuchos::rcp;
+  RCP<Tpetra::Details::ProfilingRegion> MM = rcp(new Tpetra::Details::ProfilingRegion("TpetraExt: Jacobi: Newmatrix KokkosKernels"));
 
-  // Lots and lots of typedefs
-  typedef typename Tpetra::CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>::local_matrix_host_type KCRS;
-  typedef typename KCRS::StaticCrsGraphType graph_t;
-  typedef typename graph_t::row_map_type::const_type c_lno_view_t;
-  typedef typename graph_t::entries_type::non_const_type lno_nnz_view_t;
-  typedef typename KCRS::values_type::non_const_type scalar_view_t;
-  typedef typename scalar_view_t::memory_space scalar_memory_space;
+  // Usings
+  using device_t       = typename Node::device_type;
+  using matrix_t       = typename Tpetra::CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>::local_matrix_device_type;
+  using graph_t        = typename matrix_t::StaticCrsGraphType;
+  using lno_view_t     = typename graph_t::row_map_type::non_const_type;
+  using int_view_t     = Kokkos::View<int*,
+                                  typename lno_view_t::array_layout,
+                                  typename lno_view_t::memory_space,
+                                  typename lno_view_t::memory_traits>;
+  // using c_lno_view_t   = typename graph_t::row_map_type::const_type;
+  using lno_nnz_view_t = typename graph_t::entries_type::non_const_type;
+  using scalar_view_t  = typename matrix_t::values_type::non_const_type;
 
-  typedef Scalar SC;
-  typedef LocalOrdinal LO;
-  typedef GlobalOrdinal GO;
-  typedef Node NO;
-  typedef Map<LO, GO, NO> map_type;
-  const size_t ST_INVALID = Teuchos::OrdinalTraits<LO>::invalid();
-  const LO LO_INVALID     = Teuchos::OrdinalTraits<LO>::invalid();
-  const SC SC_ZERO        = Teuchos::ScalarTraits<Scalar>::zero();
+  // KokkosKernels handle
+  using handle_t = typename KokkosKernels::Experimental::KokkosKernelsHandle<
+      typename lno_view_t::const_value_type, typename lno_nnz_view_t::const_value_type, typename scalar_view_t::const_value_type,
+      typename device_t::execution_space, typename device_t::memory_space, typename device_t::memory_space>;
 
-  // Sizes
-  RCP<const map_type> Ccolmap = C.getColMap();
-  size_t m                    = Aview.origMatrix->getLocalNumRows();
-  size_t n                    = Ccolmap->getLocalNumElements();
+  using int_handle_t = typename KokkosKernels::Experimental::KokkosKernelsHandle<
+      typename int_view_t::const_value_type, typename lno_nnz_view_t::const_value_type, typename scalar_view_t::const_value_type,
+      typename device_t::execution_space, typename device_t::memory_space, typename device_t::memory_space>;
 
-  // Grab the  Kokkos::SparseCrsMatrices & inner stuff
-  const KCRS Amat = Aview.origMatrix->getLocalMatrixHost();
-  const KCRS Bmat = Bview.origMatrix->getLocalMatrixHost();
-  const KCRS Cmat = C.getLocalMatrixHost();
+  // Merge the B and Bimport matrices
+  const matrix_t Bmerged = Tpetra::MMdetails::merge_matrices(Aview, Bview, Acol2Brow, Acol2Irow, Bcol2Ccol, Icol2Ccol, C.getColMap()->getLocalNumElements());
 
-  c_lno_view_t Arowptr = Amat.graph.row_map, Browptr = Bmat.graph.row_map, Crowptr = Cmat.graph.row_map;
-  const lno_nnz_view_t Acolind = Amat.graph.entries, Bcolind = Bmat.graph.entries, Ccolind = Cmat.graph.entries;
-  const scalar_view_t Avals = Amat.values, Bvals = Bmat.values;
-  scalar_view_t Cvals = Cmat.values;
+  // Get the properties and arrays of input matrices
+  const matrix_t Amat = Aview.origMatrix->getLocalMatrixDevice();
+  const matrix_t Bmat = Bview.origMatrix->getLocalMatrixDevice();
 
-  c_lno_view_t Irowptr;
-  lno_nnz_view_t Icolind;
-  scalar_view_t Ivals;
-  if (!Bview.importMatrix.is_null()) {
-    auto lclB = Bview.importMatrix->getLocalMatrixHost();
-    Irowptr   = lclB.graph.row_map;
-    Icolind   = lclB.graph.entries;
-    Ivals     = lclB.values;
+  typename handle_t::nnz_lno_t AnumRows = Amat.numRows();
+  typename handle_t::nnz_lno_t BnumRows = Bmerged.numRows();
+  typename handle_t::nnz_lno_t BnumCols = Bmerged.numCols();
+
+  // Arrays of the output matrix
+  lno_view_t row_mapC(Kokkos::ViewAllocateWithoutInitializing("row_mapC"), AnumRows + 1);
+  lno_nnz_view_t entriesC;
+  scalar_view_t valuesC;
+
+  std::string myalg("SPGEMM_KK_MEMORY");
+  int team_work_size;
+  std::string nodeLowerCase;
+  std::string nodeUpperCase;
+#ifdef KOKKOS_ENABLE_SERIAL
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosSerialWrapperNode>) {
+    nodeLowerCase  = "serial";
+    nodeUpperCase  = "Serial";
+    team_work_size = 1;
+  }
+#endif
+#ifdef KOKKOS_ENABLE_OPENMP
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosOpenMPWrapperNode>) {
+    nodeLowerCase  = "openmp";
+    nodeLowerCase  = "OpenMP";
+    team_work_size = 16;
+  }
+#endif
+#ifdef KOKKOS_ENABLE_CUDA
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosCudaWrapperNode>) {
+    nodeLowerCase  = "cuda";
+    nodeUpperCase  = "Cuda";
+    team_work_size = 16;
+  }
+#endif
+#ifdef KOKKOS_ENABLE_HIP
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosHipWrapperNode>) {
+    nodeLowerCase  = "hip";
+    nodeUpperCase  = "HIP";
+    team_work_size = 16;
+  }
+#endif
+#ifdef KOKKOS_ENABLE_SYCL
+  if constexpr (std::is_same_v<Node, Tpetra::KokkosCompat::KokkosSYCLWrapperNode>) {
+    nodeLowerCase  = "sycl";
+    nodeUpperCase  = "SYCL";
+    team_work_size = 16;
+  }
+#endif
+
+  // Options
+  if (!params.is_null()) {
+    if (params->isParameter(nodeLowerCase + ": algorithm"))
+      myalg = params->get(nodeLowerCase + ": algorithm", myalg);
+    if (params->isParameter(nodeLowerCase + ": team work size"))
+      team_work_size = params->get(nodeLowerCase + ": team work size", team_work_size);
   }
 
-  // Jacobi-specific inner stuff
-  auto Dvals =
-      Dinv.template getLocalView<scalar_memory_space>(Access::ReadOnly);
+  // Get the algorithm mode
+  std::string alg(nodeUpperCase + " algorithm");
+  if (!params.is_null() && params->isParameter(alg)) myalg = params->get(alg, myalg);
+  KokkosSparse::SPGEMMAlgorithm alg_enum = KokkosSparse::StringToSPGEMMAlgorithm(myalg);
 
-  // The status array will contain the index into colind where this entry was last deposited.
-  //   c_status[i] <  CSR_ip - not in the row yet
-  //   c_status[i] >= CSR_ip - this is the entry where you can find the data
-  // We start with this filled with INVALID's indicating that there are no entries yet.
-  // Sadly, this complicates the code due to the fact that size_t's are unsigned.
-  std::vector<size_t> c_status(n, ST_INVALID);
+  // decide whether to use integer-typed row pointers for this spgemm
+  Tpetra::Details::IntRowPtrHelper<decltype(Bmerged)> irph(Bmerged.nnz(), Bmerged.graph.row_map);
+  const bool useIntRowptrs =
+      irph.shouldUseIntRowptrs() &&
+      Aview.origMatrix->getApplyHelper()->shouldUseIntRowptrs();
 
-  // For each row of A/C
-  size_t CSR_ip = 0, OLD_ip = 0;
-  for (size_t i = 0; i < m; i++) {
-    // First fill the c_status array w/ locations where we're allowed to
-    // generate nonzeros for this row
-    OLD_ip = Crowptr[i];
-    CSR_ip = Crowptr[i + 1];
-    for (size_t k = OLD_ip; k < CSR_ip; k++) {
-      c_status[Ccolind[k]] = k;
+  if (useIntRowptrs) {
+    int_handle_t kh;
+    kh.create_spgemm_handle(alg_enum);
+    kh.set_team_work_size(team_work_size);
 
-      // Reset values in the row of C
-      Cvals[k] = SC_ZERO;
+    int_view_t int_row_mapC(Kokkos::ViewAllocateWithoutInitializing("int_row_mapC"), AnumRows + 1);
+
+    auto Aint = Aview.origMatrix->getApplyHelper()->getIntRowptrMatrix(Amat);
+    auto Bint = irph.getIntRowptrMatrix(Bmerged);
+
+    {
+      Tpetra::Details::ProfilingRegion MM2("TpetraExt: Jacobi: KokkosKernels symbolic int");
+      KokkosSparse::spgemm_symbolic(&kh, AnumRows, BnumRows, BnumCols,
+                                    Aint.graph.row_map, Aint.graph.entries, false,
+                                    Bint.graph.row_map, Bint.graph.entries, false,
+                                    int_row_mapC);
     }
 
-    SC minusOmegaDval = -omega * Dvals(i, 0);
+    size_t c_nnz_size = kh.get_spgemm_handle()->get_c_nnz();
+    if (c_nnz_size) {
+      entriesC = lno_nnz_view_t(Kokkos::ViewAllocateWithoutInitializing("entriesC"), c_nnz_size);
+      valuesC  = scalar_view_t(Kokkos::ViewAllocateWithoutInitializing("valuesC"), c_nnz_size);
+    }
+    Tpetra::Details::ProfilingRegion MM2("TpetraExt: Jacobi: KokkosKernels numeric int");
 
-    // Entries of B
-    for (size_t j = Browptr[i]; j < Browptr[i + 1]; j++) {
-      Scalar Bval = Bvals[j];
-      if (Bval == SC_ZERO)
-        continue;
-      LO Bij = Bcolind[j];
-      LO Cij = Bcol2Ccol[Bij];
+    // even though there is no TPL for this, we have to use the same handle that was used in the symbolic phase,
+    // so need to have a special int-typed call for this as well.
+    KokkosSparse::Experimental::spgemm_jacobi(&kh, AnumRows, BnumRows, BnumCols,
+                                              Aint.graph.row_map, Aint.graph.entries, Amat.values, false,
+                                              Bint.graph.row_map, Bint.graph.entries, Bint.values, false,
+                                              int_row_mapC, entriesC, valuesC,
+                                              omega, Dinv.getLocalViewDevice(Access::ReadOnly));
+    // transfer the integer rowptrs back to the correct rowptr type
+    Kokkos::parallel_for(
+        int_row_mapC.size(), KOKKOS_LAMBDA(int i) { row_mapC(i) = int_row_mapC(i); });
+    kh.destroy_spgemm_handle();
+  } else {
+    handle_t kh;
+    kh.create_spgemm_handle(alg_enum);
+    kh.set_team_work_size(team_work_size);
 
-      TEUCHOS_TEST_FOR_EXCEPTION(c_status[Cij] < OLD_ip || c_status[Cij] >= CSR_ip,
-                                 std::runtime_error, "Trying to insert a new entry into a static graph");
-
-      Cvals[c_status[Cij]] = Bvals[j];
+    {
+      Tpetra::Details::ProfilingRegion MM2("TpetraExt: Jacobi: KokkosKernels symbolic non-int");
+      KokkosSparse::spgemm_symbolic(&kh, AnumRows, BnumRows, BnumCols,
+                                    Amat.graph.row_map, Amat.graph.entries, false,
+                                    Bmerged.graph.row_map, Bmerged.graph.entries, false,
+                                    row_mapC);
     }
 
-    // Entries of -omega * Dinv * A * B
-    for (size_t k = Arowptr[i]; k < Arowptr[i + 1]; k++) {
-      LO Aik        = Acolind[k];
-      const SC Aval = Avals[k];
-      if (Aval == SC_ZERO)
-        continue;
-
-      if (targetMapToOrigRow[Aik] != LO_INVALID) {
-        // Local matrix
-        size_t Bk = static_cast<size_t>(targetMapToOrigRow[Aik]);
-
-        for (size_t j = Browptr[Bk]; j < Browptr[Bk + 1]; ++j) {
-          LO Bkj = Bcolind[j];
-          LO Cij = Bcol2Ccol[Bkj];
-
-          TEUCHOS_TEST_FOR_EXCEPTION(c_status[Cij] < OLD_ip || c_status[Cij] >= CSR_ip,
-                                     std::runtime_error, "Trying to insert a new entry into a static graph");
-
-          Cvals[c_status[Cij]] += minusOmegaDval * Aval * Bvals[j];
-        }
-
-      } else {
-        // Remote matrix
-        size_t Ik = static_cast<size_t>(targetMapToImportRow[Aik]);
-        for (size_t j = Irowptr[Ik]; j < Irowptr[Ik + 1]; ++j) {
-          LO Ikj = Icolind[j];
-          LO Cij = Icol2Ccol[Ikj];
-
-          TEUCHOS_TEST_FOR_EXCEPTION(c_status[Cij] < OLD_ip || c_status[Cij] >= CSR_ip,
-                                     std::runtime_error, "Trying to insert a new entry into a static graph");
-
-          Cvals[c_status[Cij]] += minusOmegaDval * Aval * Ivals[j];
-        }
-      }
+    size_t c_nnz_size = kh.get_spgemm_handle()->get_c_nnz();
+    if (c_nnz_size) {
+      entriesC = lno_nnz_view_t(Kokkos::ViewAllocateWithoutInitializing("entriesC"), c_nnz_size);
+      valuesC  = scalar_view_t(Kokkos::ViewAllocateWithoutInitializing("valuesC"), c_nnz_size);
     }
+
+    Tpetra::Details::ProfilingRegion MM2("TpetraExt: Jacobi: KokkosKernels numeric non-int");
+    KokkosSparse::Experimental::spgemm_jacobi(&kh, AnumRows, BnumRows, BnumCols,
+                                              Amat.graph.row_map, Amat.graph.entries, Amat.values, false,
+                                              Bmerged.graph.row_map, Bmerged.graph.entries, Bmerged.values, false,
+                                              row_mapC, entriesC, valuesC,
+                                              omega, Dinv.getLocalViewDevice(Access::ReadOnly));
+    kh.destroy_spgemm_handle();
   }
 
-  {
-    Tpetra::Details::ProfilingRegion MM3("TpetraExt: Jacobi: Reuse ESFC");
-    C.fillComplete(C.getDomainMap(), C.getRangeMap());
-  }
+  MM = Teuchos::null;
+  MM = rcp(new Tpetra::Details::ProfilingRegion("TpetraExt: Jacobi: Newmatrix Sort"));
+
+  // Sort & set values
+  if (params.is_null() || params->get("sort entries", true))
+    Import_Util::sortCrsEntries(row_mapC, entriesC, valuesC);
+  C.setAllValues(row_mapC, entriesC, valuesC);
+
+  MM = Teuchos::null;
+  MM = rcp(new Tpetra::Details::ProfilingRegion("TpetraExt: Jacobi: Newmatrix ESFC"));
+
+  // Final Fillcomplete
+  Teuchos::RCP<Teuchos::ParameterList> labelList = rcp(new Teuchos::ParameterList);
+  labelList->set("Timer Label", label);
+  if (!params.is_null()) labelList->set("compute global constants", params->get("compute global constants", true));
+  Teuchos::RCP<const Export<LocalOrdinal, GlobalOrdinal, Node>> dummyExport;
+  C.expertStaticFillComplete(Bview.origMatrix->getDomainMap(), Aview.origMatrix->getRangeMap(), Cimport, dummyExport, labelList);
 }
 
 /*********************************************************************************************************/
